Spawned ship into the obstacle-minigame screen and removed ship from the ForestGameScreen
