+++ source/core/src/main/com/csse3200/game/areas/SpaceGameArea.java
@@ -4,6 +4,7 @@ import com.badlogic.gdx.math.Vector2;
 import com.csse3200.game.areas.terrain.TerrainFactory;
 import com.csse3200.game.areas.terrain.TerrainFactory.TerrainType;
 import com.csse3200.game.entities.Entity;
+import com.csse3200.game.entities.factories.ShipFactory;
 import com.csse3200.game.services.ResourceService;
 import com.csse3200.game.services.ServiceLocator;
 import com.csse3200.game.components.gamearea.GameAreaDisplay;
@@ -15,9 +16,10 @@ import java.util.ArrayList;
 /** Forest area for the demo game with trees, a player, and some enemies. */
 public class SpaceGameArea extends GameArea {
     private static final Logger logger = LoggerFactory.getLogger(ForestGameArea.class);
-
+    private static final GridPoint2 SHIP_SPAWN = new GridPoint2(10, 10);
     private static final String[] spaceMiniGameTextures = {
-            "images/SpaceMiniGameBackground.png", //TODO: Replace these images with copyright free images - these are just for testing purposes!!
+            "images/SpaceMiniGameBackground.png",
+            "images/Spaceship.png"//TODO: Replace these images with copyright free images - these are just for testing purposes!!
     };
 
 
@@ -44,7 +46,7 @@ public class SpaceGameArea extends GameArea {
         displayUI();
 
         spawnTerrain();
-
+        spawnShip();
 
 
     }
@@ -71,7 +73,12 @@ public class SpaceGameArea extends GameArea {
     }
 
 
-
+    private void spawnShip()
+    {
+        Entity newShip = ShipFactory.createShip();
+        spawnEntityAt(newShip, SHIP_SPAWN, true, true);
+        targetables.add(newShip);
+    }
 
     private void loadAssets() {
         logger.debug("Loading assets");
